# [proposal] chore(webpack): improve dev build performance