--- a/config/webpack.main.js
+++ b/config/webpack.main.js
@@ -29,7 +29,11 @@ const baseConfig = {
         exclude: /node_modules/,
         use: {
           loader: "ts-loader",
-          options: { configFile: path.resolve(__dirname, "tsconfig.json") }
+          options: {
+            configFile: path.resolve(__dirname, process.env.TSCONFIG || "tsconfig.json"),
+            transpileOnly: true,
+            experimentalWatchApi: true
+          }
         }
       }
     ]