# PR Prelude

Thank you for contributing to sheikah! :)

**Please fill in this template (by putting an `x` inside
the brackets) _before_ filing your PR:**

- [X] I have read and understood [CODE_OF_CONDUCT][code] document.
- [X] I have read and understood [CONTRIBUTING][cont] document.
- [X] I have read and understood [STYLEGUIDE][style] document.
- [X] I have included tests for the changes in my PR. If not, I have included a
  rationale for why I haven't.
- [X] My code is formatted and is accepted by `yarn fmt`.
- [X] **I understand my PR may be closed if it becomes obvious I didn't
  actually perform all of these steps.**

# Why this change is necessary and useful

[Please explain **in detail** why the changes in this PR are needed.]

launching the electron app during our development workflow is taking
quite some time now. this commit introduces a couple of changes that
tells ts-loader to only transpile TypeScript. this means that whenever
you run `yarn build-dev` or `yarn dev` the project will be compiled
BUT NOT type checked (for production builds though the type check step
does happen). in order to type check the project we should run `yarn
typecheck [--watch]` or `yarn typecheck-test [--watch]` separately

before this change `yarn build-dev` reports in my computer: `85.29s user 2.25s system 121% cpu 1:12.24 total`
after this change `yarn build-dev` reports in my computer: `15.10s user 0.65s system 162% cpu 9.706 total`
[code]: https://github.com/witnet/sheikah/blob/master/.github/CODE_OF_CONDUCT.md
[cont]: https://github.com/witnet/sheikah/blob/master/.github/CONTRIBUTING.md
[style]: https://github.com/witnet/sheikah/blob/master/docs/STYLEGUIDE.md